+++ NEWS.md
@@ -1,6 +1,11 @@
 # jsmodule 1.1.1
 
+## Update
+
 * Add landmark analysis to kaplan-meier analysis. Because it uses **jskm 0.4.0** version, please install it with `remotes::install_github("jinseob2kim/jskm")`
+
+* Add option **hide 95%CI of AUC, NRIIDI** to `timerocModule`. It reduces computation time.NRIIID
+
 
 # jsmodule 1.1.0